requirements.txt
Django==1.7.1
Markdown==2.5.1
Pillow==2.6.1
Shapely==1.4.4
argparse==1.2.2
boto==2.34.0
cryptography==1.4
dateutils==0.6.6
django-cors-headers==0.13
django-crum==0.6.1
django-debug-toolbar==1.4
django-extensions==1.5.5
django-filter==0.8
django-haystack==2.3.1
django-hstore==1.3.5
django-leaflet==0.15.0
django-redis==3.8.0
django-taggit==0.17.0
django-treebeard==2.0
djangorestframework==2.4.3
drf-extensions==0.2.6
elasticsearch==1.2.0
facebook-sdk==0.4.0
gdata==2.0.18
google-api-python-client==1.4.1
mock==1.0.1
mockredispy==2.9.0.11
oauth2client==1.4.11
pika==0.9.14
django-test-without-migrations==0.2
-e git+https://github.com/pphetra/pyth.git#egg=pyth-0.7.0beta
-e git+https://bitbucket.org/opendream/podd-form-generator#egg=podd_form_generator
psycopg2==2.5.4
py2neo==3
pyOpenSSL==0.15.1
python-dateutil==2.2
pytz==2014.9
redis==2.10.3
six==1.8.0
unicodecsv==0.9.4
urllib3==1.9.1
wsgiref==0.1.2
xlwt==0.7.5
celery==3.1.17
django-mailgun==0.2.2
raven==5.1.1
xlrd==0.9.4
uuid==1.30
freezegun==0.3.5
asteval==0.9.5
django-cacheops==2.4.1
opbeat==3.2
